move code editor and markdown editor
