+++ frontend/app/src/entities/proposed-changes/ui/create-form.tsx
@@ -9,7 +9,7 @@ import { useSchema } from "@/entities/schema/hooks/useSchema";
 import { useMutation } from "@/shared/api/graphql/useQuery";
 import { constructPath } from "@/shared/api/rest/fetch";
 import { LinkButton } from "@/shared/components/buttons/button-primitive";
-import { MarkdownEditor } from "@/shared/components/editor";
+import { MarkdownEditor } from "@/shared/components/editor/markdown";
 import { RelationshipManyInput } from "@/shared/components/inputs/relationship-many";
 import { ALERT_TYPES, Alert } from "@/shared/components/ui/alert";
 import { Card } from "@/shared/components/ui/card";
