+++ frontend/app/src/shared/components/conversations/comment.tsx
@@ -1,6 +1,6 @@
 import { Avatar } from "@/shared/components/display/avatar";
 import { DateDisplay } from "@/shared/components/display/date-display";
-import { MarkdownViewer } from "@/shared/components/editor/markdown-viewer";
+import { MarkdownViewer } from "@/shared/components/editor/markdown/markdown-viewer";
 import { classNames } from "@/shared/utils/common";
 import React from "react";
