+++ frontend/app/src/shared/components/editor/code/code-viewer-limiter.tsx
